+++ pybop/problems/problem.py
@@ -270,7 +270,7 @@ class Problem:
         cost0 = np.abs(self.evaluate(x0).values[0])
         nsamples = 0
         while np.isinf(cost0) and nsamples < 10:
-            x0 = self.parameters.sample_from_distributions()[0]
+            x0 = self.parameters.sample_from_distribution()[0]
             if x0 is None:
                 break