+++ tests/unit/test_posterior.py
@@ -111,8 +111,8 @@ class TestLogPosterior:
         assert problem._cost.joint_prior is not None
         assert isinstance(problem._cost.joint_prior, pybop.JointDistribution)
 
-        for i, p in enumerate(problem.parameters.distributions()):
-            assert p == problem._cost.joint_prior._distributions[i]
+        for i, p in enumerate(problem.parameters.distribution()):
+            assert p == problem._cost.joint_prior._distribution[i]
 
     @pytest.fixture
     def problem(self, simulator, likelihood, prior):